+++ app/controllers/plate_controller.py
@@ -1,10 +1,66 @@
-from app.service.plate_service import update_user_platestoverified, get_public_plates_notUser, create_plate, get_user_plates, delete_Plate_service, update_Plate, getPlateByID, get_public_plates
+from app.service.plate_service import get_plates, update_user_platestoverified, get_public_plates_notUser, create_plate, get_user_plates, delete_Plate_service, update_Plate, getPlateByID, get_public_plates
 from app.models.plate import Plate
 from fastapi import HTTPException
 from datetime import datetime
+from app.controllers.food_controller import get_foods
+import re
+
+
+def validate_name(campo, label):
+    if not campo.strip():
+        raise HTTPException(
+            status_code=400, detail=f"{label} cannot be empty or blank")
+    if type(campo) is not str:
+        raise HTTPException(
+            status_code=400, detail=f"{label} must be a string")
+    if not re.match(r'^[a-zA-Z\s]+$', campo):
+        raise HTTPException(
+            status_code=400, detail=f"Invalid format for {label}. Only letters and spaces allowed.")
+
+
+def validate_limit(campo, minimo, label):
+    if type(campo) is not float:
+        raise HTTPException(
+            status_code=400, detail=f"{label} must be a float")
+    if not (minimo <= campo):
+        raise HTTPException(
+            status_code=400, detail=f"{label} must be grater than {minimo}")
+
+
+def validate_verified(campo, minimo, label):
+    if type(campo) is not int:
+        raise HTTPException(
+            status_code=400, detail=f"{label} must be an intenger")
+    if not (minimo <= campo):
+        raise HTTPException(
+            status_code=400, detail=f"{label} must be grater than {minimo}")
+
+
+def validate_plate_data(plate: Plate):
+    validate_name(plate.name, "name")
+    validate_limit(plate.calories_portion, 0, "calories_portion")
+    validate_limit(plate.sodium_portion, 0, "sodium_portion")
+    validate_limit(plate.carbohydrates_portion, 0, "carbohydrates_portion")
+    validate_limit(plate.fats_portion, 0, "fats_portion")
+    validate_limit(plate.protein_portion, 0, "protein_portion")
+    validate_verified(plate.verified, 0, "verified")
+    if not plate.ingredients:
+        raise HTTPException(
+            status_code=400, detail=f"The plate must to have at least one ingredient")
+
+    foods = get_foods()
+    existing_food_ids = [food["id"]
+                         for food in foods["message"]['food']]
+
+    for ing in plate.ingredients:
+        validate_limit(ing.quantity, 0, f"quantity of {ing.ingredientId}")
+        if ing.ingredientId not in existing_food_ids:
+            raise HTTPException(
+                status_code=400, detail=f"Invalid food ID: {ing.ingredientId} ")
 
 
 def plateLog(plate: Plate):
+    validate_plate_data(plate)
     response = create_plate(plate)
     if "error" in response:
         raise HTTPException(status_code=500, detail=response["error"])
@@ -18,14 +74,33 @@ def get_plate_user(user_id: str):
     return {"message": response}
 
 
-def delete_plate(userPlate_id: str):
+def delete_plate(user_id: str, userPlate_id: str):
+    user_plates = get_plate_user(user_id)
+    existing_plates_ids = [plate["id"]
+                           for plate in user_plates["message"]['Plates']]
+    if userPlate_id not in existing_plates_ids:
+        raise HTTPException(
+            status_code=400, detail=f"The plate Id is not valid")
     response = delete_Plate_service(userPlate_id)
     if "error" in response:
         raise HTTPException(status_code=500, detail=response["error"])
     return response
 
 
-def update_plate_controller(userPlate_id: str, Plate_data: Plate):
+def update_plate_controller(id_user: str, userPlate_id: str, Plate_data: Plate):
+    user_plates = get_plate_user(id_user)
+    existing_plates_ids = [plate["id"]
+                           for plate in user_plates["message"]['Plates']]
+    if userPlate_id not in existing_plates_ids:
+        raise HTTPException(
+            status_code=400, detail=f"The plate Id is not valid")
+
+    if id_user != Plate_data.id_User:
+        raise HTTPException(
+            status_code=400, detail=f"You can not change id_User")
+
+    validate_plate_data(Plate_data)
+
     response = update_Plate(userPlate_id, Plate_data)
     if "error" in response:
         raise HTTPException(status_code=500, detail=response["error"])
@@ -33,7 +108,15 @@ def update_plate_controller(userPlate_id: str, Plate_data: Plate):
 
 
 def get_platebyID(plate_id: str):
+    plates = get_plates()
+    existing_plate_ids = [plate["id"] for plate in plates]
+    if plate_id not in existing_plate_ids:
+        raise HTTPException(
+            status_code=400, detail=f"Invalid plate ID: {plate_id}")
     response = getPlateByID(plate_id)
+    if response['plate'] is None:
+        raise HTTPException(
+            status_code=400, detail=f"The plate ID: {plate_id} is not in our data base")
     if "error" in response:
         raise HTTPException(status_code=500, detail=response["error"])
     return {"message": response}
